requirements.txt
absl-py==0.1.10
jpype1>=0.7.0
